Rename mdx files
