+++ components/mdx/index.tsx
@@ -2,10 +2,10 @@ export { P } from "./p";
 export { H1 } from "./h1";
 // export { H2 } from "./h2";
 export { H3 } from "./h3";
-export { Callout } from "./callout2";
+export { Callout } from "./callout";
 export { CodeSnippets } from "./code-snippets";
 export { CustomImage } from "./custom-image";
 export { FeatureSnippets } from "./feature-snippets";
 export { IfRenderer } from "./if-renderer";
-export { Video } from "./video2";
+export { Video } from "./video";
 export { YouTubeCallout } from "./youtube-callout";
